requirements.txt
python-dotenv>=1.0.1
longport>=2.1.10
requests>=2.32.5
pandas-ta>=0.4.71b0
openpyxl>=3.1.5
duckdb>=1.4.2
